requirements.txt
# ============================================================
# Seeing Through Touch - Python Dependencies
# ============================================================
# NOTE: PyTorch, torchvision, torchaudio, and KNN_CUDA are NOT
# included here. Install them separately (see README.md).
# ============================================================

# --- Core ML / Vision ---
timm>=1.0
open-clip-torch>=2.24
torchmetrics>=1.0
einops>=0.8
xformers==0.0.31  # compatible with torch 2.7.1; update alongside torch if using a newer version
fairscale>=0.4

# --- Scientific computing ---
numpy>=1.26,<2.0
scipy>=1.10
scikit-learn>=1.2
pandas>=2.0
numba>=0.60

# --- Image / Video ---
pillow>=10.0
opencv-python>=4.8
pycocotools>=2.0
open3d>=0.18
ImageHash>=4.3

# --- Visualization ---
matplotlib>=3.7
seaborn>=0.13
plotly>=5.0
dash>=3.0

# --- Config & CLI ---
hydra-core>=1.3
omegaconf>=2.3
pyyaml>=6.0
fire>=0.5
easydict>=1.13
ConfigArgParse>=1.7

# --- Logging & Experiment tracking ---
wandb>=0.16
tensorboard>=2.14
rich>=13.0
tqdm>=4.65

# --- HuggingFace ecosystem ---
transformers>=4.40
accelerate>=1.0
peft>=0.10
diffusers>=0.28
safetensors>=0.4
huggingface-hub>=0.20
sentencepiece>=0.2
tokenizers>=0.21

# --- Retrieval ---
faiss-cpu>=1.7
autofaiss>=2.17

# --- Data I/O ---
h5py>=3.10
pyarrow>=15.0

# --- Misc ---
pyquaternion>=0.9
gradio>=5.0
pygame>=2.6
addict>=2.4
antlr4-python3-runtime>=4.9,<4.10
